move PercentEncoder to it's own space
