+++ core/shared/src/main/scala/uritemplate4s/UriTemplate.scala
@@ -1,6 +1,6 @@
 package uritemplate4s
 
-import fastparse.all._
+import fastparse.all.Parsed
 import uritemplate4s.Error.MalformedUriTemplate
 import uritemplate4s.ListSyntax._
 import uritemplate4s.UriTemplate._
@@ -244,17 +244,3 @@ private sealed trait ModifierLevel4
 private case object EmptyModifier extends ModifierLevel4
 private final case class Prefix(maxLength: Int) extends ModifierLevel4
 private case object Explode extends ModifierLevel4
-
-private object PercentEncoder {
-  import UriTemplateParser._
-
-  @inline def percentEncode(s: String): String =
-    s.getBytes("UTF-8").flatMap(byte => encodeChar(byte.toChar)).mkString
-
-  private def encodeChar(ch: Char): String = s"%${"%04x".format(ch.toInt).substring(2).toUpperCase}"
-
-  lazy val nonUnreserved: P[List[Literals]] =
-    P(unreserved.rep(min = 1).!.map(Encoded) | (!unreserved ~ AnyChar).rep(min = 1).!.map(Unencoded)).rep.map(_.toList)
-  lazy val nonUnreservedAndReserved: P[List[Literals]] =
-    P((unreserved | reserved).rep(min = 1).!.map(Encoded) | (!(unreserved | reserved) ~ AnyChar).rep(min = 1).!.map(Unencoded)).rep.map(_.toList)
-}
